Refactor LeadDetailScreen by extracting TabbedTimelineCard to a separate file
- Moved the TabbedTimelineCard implementation to timeline/tabbed_timeline_card.dart for better code organization and maintainability.
- Updated references in LeadDetailScreen to use the new TabbedTimelineCard component.
- Corrected database table names in Supabase queries for consistency.
